+++ src/test/java/domain/piece/KingTest.java
@@ -1,25 +1,24 @@
-package domain;
+package domain.piece;
 
 import static domain.PieceMoveResult.CATCH;
 import static domain.PieceMoveResult.CATCH_KING;
 import static domain.PieceMoveResult.FAILURE;
 import static domain.PieceMoveResult.SUCCESS;
 import static domain.Position.A4;
-import static domain.Position.B3;
-import static domain.Position.B5;
-import static domain.Position.C2;
-import static domain.Position.C6;
+import static domain.Position.C3;
+import static domain.Position.C4;
+import static domain.Position.C5;
+import static domain.Position.D3;
 import static domain.Position.D4;
-import static domain.Position.E2;
-import static domain.Position.E6;
-import static domain.Position.F3;
-import static domain.Position.F5;
+import static domain.Position.D5;
+import static domain.Position.E3;
+import static domain.Position.E4;
+import static domain.Position.E5;
 import static domain.Team.BLACK;
 import static domain.Team.WHITE;
 
-import domain.piece.King;
-import domain.piece.Knight;
-import domain.piece.Pawn;
+import domain.PiecesOnChessBoard;
+import domain.Position;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Set;
@@ -32,7 +31,7 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
-class KnightTest {
+class KingTest {
     public static Stream<Arguments> moveFailureParameters() {
         Set<Position> successParameters = moveSuccessParameters().map(Arguments::get)
                 .map(objects -> (Position) objects[0])
@@ -44,68 +43,67 @@ class KnightTest {
 
     public static Stream<Arguments> moveSuccessParameters() {
         return Stream.of(
-                Arguments.of(C6), Arguments.of(E6),
-                Arguments.of(B5), Arguments.of(F5),
-                Arguments.of(B3), Arguments.of(F3),
-                Arguments.of(C2), Arguments.of(E2)
+                Arguments.of(C5), Arguments.of(D5), Arguments.of(E5),
+                Arguments.of(C4), Arguments.of(E4),
+                Arguments.of(C3), Arguments.of(D3), Arguments.of(E3)
         );
     }
 
     @ParameterizedTest
     @MethodSource("moveSuccessParameters")
-    @DisplayName("나이트의 이동 규칙대로 이동이 가능한지 검증")
+    @DisplayName("킹의 이동 규칙대로 이동이 가능한지 검증")
     void moveSuccess(Position targetPosition) {
-        Knight knight = new Knight(D4, WHITE);
+        King king = new King(D4, WHITE);
         PiecesOnChessBoard piecesOnChessBoard = new PiecesOnChessBoard(List.of());
-        Assertions.assertThat(knight.move(targetPosition, piecesOnChessBoard))
+        Assertions.assertThat(king.move(targetPosition, piecesOnChessBoard))
                 .isEqualTo(SUCCESS);
     }
 
     @ParameterizedTest
     @MethodSource("moveFailureParameters")
-    @DisplayName("나이트의 이동 규칙을 위반한 이동이 불가능한지 검증")
+    @DisplayName("킹의 이동 규칙을 위반한 이동이 불가능한지 검증")
     void moveFailure(Position targetPosition) {
-        Knight knight = new Knight(D4, WHITE);
+        King king = new King(D4, WHITE);
         PiecesOnChessBoard piecesOnChessBoard = new PiecesOnChessBoard(List.of());
-        Assertions.assertThat(knight.move(targetPosition, piecesOnChessBoard))
+        Assertions.assertThat(king.move(targetPosition, piecesOnChessBoard))
                 .isEqualTo(FAILURE);
     }
 
     @ParameterizedTest
     @MethodSource("moveSuccessParameters")
-    @DisplayName("나이트의 목적지에 같은 팀 말이 있는 경우 이동이 불가능한지 검증")
+    @DisplayName("킹의 목적지에 같은 팀 말이 있는 경우 이동이 불가능한지 검증")
     void moveFailureCauseTargetIsSameTeam(Position targetPosition) {
-        Knight knight = new Knight(D4, WHITE);
+        King king = new King(D4, WHITE);
         PiecesOnChessBoard piecesOnChessBoard = new PiecesOnChessBoard(List.of(new Pawn(targetPosition, WHITE)));
-        Assertions.assertThat(knight.move(targetPosition, piecesOnChessBoard))
+        Assertions.assertThat(king.move(targetPosition, piecesOnChessBoard))
                 .isEqualTo(FAILURE);
     }
 
     @ParameterizedTest
     @MethodSource("moveSuccessParameters")
-    @DisplayName("나이트의 목적지에 다른 팀 말이 있는 경우 이동이 가능한지 검증")
+    @DisplayName("킹의 목적지에 다른 팀 말이 있는 경우 이동이 가능한지 검증")
     void moveSuccessWhenTargetIsOtherTeam(Position targetPosition) {
-        Knight knight = new Knight(D4, WHITE);
+        King king = new King(D4, WHITE);
         PiecesOnChessBoard piecesOnChessBoard = new PiecesOnChessBoard(List.of(new Pawn(targetPosition, BLACK)));
-        Assertions.assertThat(knight.move(targetPosition, piecesOnChessBoard))
+        Assertions.assertThat(king.move(targetPosition, piecesOnChessBoard))
                 .isEqualTo(CATCH);
     }
 
     @Test
-    @DisplayName("나이트의 목적지에 다른 팀 킹이 있는 경우 CATCH_KING을 반환하는지 검증")
+    @DisplayName("킹의 목적지에 다른 팀 킹이 있는 경우 CATCH_KING을 반환하는지 검증")
     void catchKing() {
-        Knight knight = new Knight(D4, WHITE);
-        PiecesOnChessBoard piecesOnChessBoard = new PiecesOnChessBoard(List.of(new King(C6, BLACK)));
-        Assertions.assertThat(knight.move(C6, piecesOnChessBoard))
+        King king = new King(D4, WHITE);
+        PiecesOnChessBoard piecesOnChessBoard = new PiecesOnChessBoard(List.of(new King(D5, BLACK)));
+        Assertions.assertThat(king.move(D5, piecesOnChessBoard))
                 .isEqualTo(CATCH_KING);
     }
 
     @Test
-    @DisplayName("나이트의 목적지에 같은 팀 킹이 있는 경우 이동이 불가능한지 검증")
+    @DisplayName("킹의 목적지에 같은 팀 킹이 있는 경우 이동이 불가능한지 검증")
     void moveFailureByOurKing() {
-        Knight knight = new Knight(D4, WHITE);
-        PiecesOnChessBoard piecesOnChessBoard = new PiecesOnChessBoard(List.of(new King(C6, WHITE)));
-        Assertions.assertThat(knight.move(A4, piecesOnChessBoard))
+        King king = new King(D4, WHITE);
+        PiecesOnChessBoard piecesOnChessBoard = new PiecesOnChessBoard(List.of(new King(D5, WHITE)));
+        Assertions.assertThat(king.move(A4, piecesOnChessBoard))
                 .isEqualTo(FAILURE);
     }
 }
