+++ src/test/java/domain/piece/PawnTest.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.piece;
 
 import static domain.PieceMoveResult.CATCH;
 import static domain.PieceMoveResult.CATCH_KING;
@@ -19,8 +19,10 @@ import static domain.Position.E5;
 import static domain.Team.BLACK;
 import static domain.Team.WHITE;
 
-import domain.piece.King;
-import domain.piece.Pawn;
+import domain.PieceMoveResult;
+import domain.PiecesOnChessBoard;
+import domain.Position;
+import domain.Team;
 import java.util.List;
 import java.util.stream.Stream;
 import org.assertj.core.api.Assertions;
