+++ src/test/java/domain/piece/RookTest.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.piece;
 
 import static domain.PieceMoveResult.CATCH;
 import static domain.PieceMoveResult.CATCH_KING;
@@ -22,9 +22,8 @@ import static domain.Position.H4;
 import static domain.Team.BLACK;
 import static domain.Team.WHITE;
 
-import domain.piece.King;
-import domain.piece.Pawn;
-import domain.piece.Rook;
+import domain.PiecesOnChessBoard;
+import domain.Position;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Set;
